Удаление предупреждений PVS-Studio

+++ source/TacticsVictory/src/Core/Math.cpp
@@ -18,7 +18,7 @@ BoundingBox Math::CalculateBoundingBox(float *buffer, uint numVertexes)
 
     for(uint i = 0; i < numVertexes; i++)
     {
-        box.Merge(Vector3(buffer + (size_t)(i * 8)));
+        box.Merge(Vector3(buffer + (uint64)(i * 8)));
     }
 
     return box;
